Update: Add some FailFast code

+++ HairSaloonApp/HairSaloonAPI/Services/UserService.cs
@@ -4,6 +4,7 @@ using HairSaloonAPI.Interfaces.Services;
 using HairSaloonAPI.Models;
 using HairSaloonAPI.Models.DTOs.ControllerDTOs;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.IdentityModel.Tokens;
 
 namespace HairSaloonAPI.Services;
 
@@ -18,6 +19,10 @@ public class UserService : IUserService
 
     public void DeleteUser(int id)
     {
+        if (!_db.Users.Any(u => u.Id == id))
+        {
+            throw new BadHttpRequestException("No User was found");
+        }
         User _user = _db.Users.First(u => u.Id == id);
 
         _db.Users.Remove(_user);
@@ -26,12 +31,20 @@ public class UserService : IUserService
 
     public int GetLastUserId()
     {
+        if (_db.Users.IsNullOrEmpty())
+        {
+            throw new BadHttpRequestException("No User in the Database");
+        }
         User _lastUser = _db.Users.OrderByDescending(u => u.Id).First();
         return _lastUser.Id;
     }
 
     public List<UserListDTO> GetAllUsers()
     {
+        if (_db.Persons.IsNullOrEmpty())
+        {
+            throw new BadHttpRequestException("No User in the Database");
+        }
         List<Person> dbUserList = _db.Persons.Include(p => p.User).ToList();
         List<UserListDTO> users = new List<UserListDTO>();
         dbUserList.ForEach(p => users.Add(new UserListDTO(p.FirstName, p.LastName, p.User.Id)));
@@ -41,6 +54,10 @@ public class UserService : IUserService
 
     public List<UserListDTO> GetAllHairDressers()
     {
+        if (_db.Persons.IsNullOrEmpty())
+        {
+            throw new BadHttpRequestException("No User in the Database");
+        }
         List<Person> dbUserList = _db.Persons.Include(p => p.User).Where(p => p.Type == PersonType.HairDresser).ToList();
         List<UserListDTO> users = new List<UserListDTO>();
         dbUserList.ForEach(p => users.Add(new UserListDTO(p.FirstName, p.LastName, p.User.Id)));
